+++ src/main/java/taxiservice/controller/car/CreateCarController.java
@@ -1,4 +1,4 @@
-package taxiservice.controllers.manufacturers;
+package taxiservice.controller.car;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,26 +6,29 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import taxiservice.lib.Injector;
-import taxiservice.model.Manufacturer;
+import taxiservice.model.Car;
+import taxiservice.service.CarService;
 import taxiservice.service.ManufacturerService;
 
-public class ManufacturerCreateController extends HttpServlet {
+public class CreateCarController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("taxiservice");
     private static ManufacturerService manufacturerService = (ManufacturerService)
             injector.getInstance(ManufacturerService.class);
+    private static CarService carService = (CarService)
+            injector.getInstance(CarService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/views/manufacturers/add.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/car/add.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Manufacturer input = new Manufacturer(req.getParameter("name"),
-                req.getParameter("country"));
-        manufacturerService.create(input);
-        resp.sendRedirect(req.getContextPath() + "/manufacturers/all");
+        Car input = new Car(req.getParameter("model"),
+                manufacturerService.get(Long.valueOf(req.getParameter("manufacturer_id"))));
+        carService.create(input);
+        resp.sendRedirect(req.getContextPath() + "/cars/all");
     }
 }
