+++ src/main/java/taxiservice/controller/car/driver/AddDriverToCarController.java
@@ -1,4 +1,4 @@
-package taxiservice.controllers.drivers;
+package taxiservice.controller.car.driver;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,25 +6,27 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import taxiservice.lib.Injector;
-import taxiservice.model.Driver;
+import taxiservice.service.CarService;
 import taxiservice.service.DriverService;
 
-public class DriverCreateController extends HttpServlet {
+public class AddDriverToCarController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("taxiservice");
+    private static CarService carService = (CarService)
+            injector.getInstance(CarService.class);
     private static DriverService driverService = (DriverService)
             injector.getInstance(DriverService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/views/drivers/add.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/car/driver/add.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Driver input = new Driver(req.getParameter("name"), req.getParameter("licence_number"));
-        driverService.create(input);
-        resp.sendRedirect(req.getContextPath() + "/drivers/all");
+        carService.addDriverToCar(driverService.get(Long.valueOf(req.getParameter("driver_id"))),
+                carService.get(Long.valueOf(req.getParameter("car_id"))));
+        resp.sendRedirect(req.getContextPath() + "/cars/all");
     }
 }
