+++ src/main/java/taxiservice/controller/manufacturer/CreateManufacturerController.java
@@ -1,4 +1,4 @@
-package taxiservice.controllers.cars;
+package taxiservice.controller.manufacturer;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,29 +6,26 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import taxiservice.lib.Injector;
-import taxiservice.model.Car;
-import taxiservice.service.CarService;
+import taxiservice.model.Manufacturer;
 import taxiservice.service.ManufacturerService;
 
-public class CarCreateController extends HttpServlet {
+public class CreateManufacturerController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("taxiservice");
     private static ManufacturerService manufacturerService = (ManufacturerService)
             injector.getInstance(ManufacturerService.class);
-    private static CarService carService = (CarService)
-            injector.getInstance(CarService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/views/cars/add.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/manufacturer/add.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Car input = new Car(req.getParameter("model"),
-                manufacturerService.get(Long.valueOf(req.getParameter("manufacturer_id"))));
-        carService.create(input);
-        resp.sendRedirect(req.getContextPath() + "/cars/all");
+        Manufacturer input = new Manufacturer(req.getParameter("name"),
+                req.getParameter("country"));
+        manufacturerService.create(input);
+        resp.sendRedirect(req.getContextPath() + "/manufacturers/all");
     }
 }
